[Adjustment][Antonio] Code cleanup
